# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/request": {
            "post": {
                "tags": [
                    "Request"
                ],
                "summary": "Создание заявки.",
                "operationId": "make_request_api_request_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Заявка была успешно создана.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RequestCreateResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Заявка уже была подана."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/request/{request_id}": {
            "get": {
                "tags": [
                    "Request"
                ],
                "summary": "Получение заявки по её ID.",
                "operationId": "get_request_by_id_api_request__request_id__get",
                "parameters": [
                    {
                        "name": "request_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Request Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заявка успешно получена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Request"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заявка не найдена."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Request"
                ],
                "summary": "Изменение заявки",
                "operationId": "patch_request_api_request__request_id__patch",
                "parameters": [
                    {
                        "name": "request_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Request Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestPatch"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Заявка была успешно изменена.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Request"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заявка не была найдена"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/request/student/{student_id}": {
            "get": {
                "tags": [
                    "Request"
                ],
                "summary": "Получение всех заявок по студента по его ID.",
                "operationId": "get_requests_by_student_id_api_request_student__student_id__get",
                "parameters": [
                    {
                        "name": "student_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Student Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заявки были успешно получены.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Request"
                                    },
                                    "title": "Response 200 Get Requests By Student Id Api Request Student  Student Id  Get"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заявка не была найдена"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/request/mentor/{mentor_id}": {
            "get": {
                "tags": [
                    "Request"
                ],
                "summary": "Получение всех заявок ментора",
                "operationId": "get_requests_by_mentor_id_api_request_mentor__mentor_id__get",
                "parameters": [
                    {
                        "name": "mentor_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Mentor Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заявки были успешно получены.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Request"
                                    },
                                    "title": "Response 200 Get Requests By Mentor Id Api Request Mentor  Mentor Id  Get"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заявка не была найдена"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/stats/{mentor_id}": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "Получение статистики по ментору.",
                "operationId": "get_mentor_stats_api_stats__mentor_id__get",
                "parameters": [
                    {
                        "name": "mentor_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Mentor Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статистика была успешно получена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MentorStats"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Статистика не была найдена."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/stats": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "Получение статистики по всему серверу",
                "operationId": "get_stats_api_stats_get",
                "responses": {
                    "200": {
                        "description": "Статистика была успешно получена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminStats"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Статистика не была найдена."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdminStats": {
                "properties": {
                    "total_students": {
                        "type": "integer",
                        "title": "Total Students",
                        "description": "Полное количество студентов"
                    },
                    "total_mentors": {
                        "type": "integer",
                        "title": "Total Mentors",
                        "description": "Полное количество менторов"
                    },
                    "accepted": {
                        "type": "integer",
                        "title": "Accepted",
                        "description": "Полное количество одобренных заявок"
                    },
                    "refused": {
                        "type": "integer",
                        "title": "Refused",
                        "description": "Полное количество отклоненных заявок"
                    },
                    "pending": {
                        "type": "integer",
                        "title": "Pending",
                        "description": "Количество ожидающих заявок"
                    }
                },
                "type": "object",
                "required": [
                    "total_students",
                    "total_mentors",
                    "accepted",
                    "refused",
                    "pending"
                ],
                "title": "AdminStats"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "MentorStats": {
                "properties": {
                    "accepted": {
                        "type": "integer",
                        "title": "Accepted",
                        "description": "Сколько заявок было принято"
                    },
                    "refused": {
                        "type": "integer",
                        "title": "Refused",
                        "description": "Сколько заявок было отклнонено"
                    },
                    "pending": {
                        "type": "integer",
                        "title": "Pending",
                        "description": "Сколько заявок ожидает"
                    },
                    "total": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Total",
                        "description": "Сколько всего заявок",
                        "default": 0
                    }
                },
                "type": "object",
                "required": [
                    "accepted",
                    "refused",
                    "pending"
                ],
                "title": "MentorStats"
            },
            "Request": {
                "properties": {
                    "request_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Request Id",
                        "description": "ID заявки"
                    },
                    "mentor_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Mentor Id",
                        "description": "ID ментора"
                    },
                    "student_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Student Id",
                        "description": "ID клиента"
                    },
                    "query": {
                        "type": "string",
                        "title": "Query",
                        "description": "Вопрос пользователя"
                    },
                    "status": {
                        "$ref": "#/components/schemas/Status",
                        "description": "Статус заявки pending / accepted / refused",
                        "default": "PENDING"
                    },
                    "date": {
                        "type": "string",
                        "format": "yy-mm-dd",
                        "title": "Date",
                        "description": "Дата в формате YY-MM-DD"
                    }
                },
                "type": "object",
                "required": [
                    "request_id",
                    "mentor_id",
                    "student_id",
                    "query",
                    "date"
                ],
                "title": "Request"
            },
            "RequestCreate": {
                "properties": {
                    "mentor_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Mentor Id",
                        "description": "ID ментора"
                    },
                    "student_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Student Id",
                        "description": "ID клиента"
                    },
                    "query": {
                        "type": "string",
                        "title": "Query",
                        "description": "Вопрос пользователя"
                    },
                    "status": {
                        "$ref": "#/components/schemas/Status",
                        "description": "Статус заявки pending / accepted / refused",
                        "default": "PENDING"
                    },
                    "date": {
                        "type": "string",
                        "format": "yy-mm-dd",
                        "title": "Date",
                        "description": "Дата в формате YY-MM-DD"
                    }
                },
                "type": "object",
                "required": [
                    "mentor_id",
                    "student_id",
                    "query",
                    "date"
                ],
                "title": "RequestCreate"
            },
            "RequestCreateResponse": {
                "properties": {
                    "request_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Request Id",
                        "description": "ID заявки"
                    }
                },
                "type": "object",
                "required": [
                    "request_id"
                ],
                "title": "RequestCreateResponse"
            },
            "RequestPatch": {
                "properties": {
                    "query": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Query",
                        "description": "Запрос юзера"
                    },
                    "status": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Status",
                        "description": "Статус заявки"
                    },
                    "date": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "format": "yy-mm-dd",
                        "title": "Date",
                        "description": "Дата в формате YY-MM-DD"
                    }
                },
                "type": "object",
                "title": "RequestPatch"
            },
...
            "Status": {
                "type": "string",
                "enum": [
                    "ACCEPTED",
                    "REFUSED",
                    "PENDING"
                ],
                "title": "Status"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "CustomHTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}